AI-936 Unused file cleanup.
Have removed the JsonRpcClientTest and friends- from here on out, let's put any tests that need to touch
the database in the acceptance test suite so we can run it against the deployed server.

Otherwise we have many tests that are not catching problems - for instance, there was  fix for date
formatting that was essentially added to JsonRpcTest, but masked the fact the server was still
returning milliseconds since the epoch instead of YYYY-MM-dd, which would affect those using the api
outside of our test suite.