AI-304 Updated capacity test to verify sync functionality under realistic load
This bug has remained open for a very long time because it only occurs when the system
is under a relatively high load. The test case recently written for this bug, for example,
actually passes on the current production version if run in isolation.

This updated capacity test mimics a realistic load by using a very crude version of
agent-based simulations: we create a basic model of a scenario that includes ten sectors,
each with their own database, and 150 partner organizations, each with 10 users.

The test is broken up into 'days', during which different agents are programmed
to carry out different tasks against the server.

For example, the sector leads are programmed to each set up their database on day 0,
and the reporting users start synchronizing their local databases beginning on day 1.
The sector leads continuing updating the structure while the reprorting users synchronize.

Under this load, the current version (not this branch) begins to fail with 50+ concurrent
users and synchronization payloads > 800k.