AI-600 Improved synchronization simulation and increased load
Previously, we were load testing number of concurrent *complete* synchronizations, which was
useful for sussing out performance problems in the synchronization mechanism, but was time-consuming
and difficult to get a realistic estimate of 'engineered capacity'.

This commit updates the sync mechanism to work as it does in the real client, where only
updated sync regions are fetched by our modelled army of users.