Improves robustness of both application and acceptance tests under poor network conditions
First, this commit fixes the behavior of the pivot table editor (and other reports) under poor network
conditions that lead to updates arriving out of order.

This commit also increases the timeout of automated UI tests to well beyond 60 seconds to avoid failing
a test that is adversely affected by a random spike in latency caused by factors  exogenous to the system under
test.

As the number of tests increase, the probabilty of a single test failing due to such spikes increases and
as long as the application recovers, the time required to run the test should not be counted as a failure
but rather evidence of robustness under realistic conditions.

To this end, this commit also introduces the ProxyController, something I've long wanted to implement.

The ProxyController sits between the WebDriver-controlled browser and the server, and through the use
of filters, can simulate adverse network connections such as high latency typical of a satellite connection,
or unreliability characteristic of the poor connections upon which many of users rely.