AI-822 Changed the strategy for matching source/target rows to one based on pareto optimal matches
Previously, source/target matching was solved using the solution to the 'assignment problem', which
views the problem as choosing the best matchings between two sets of objects by optimizing the 'cost'
of those pairings.

This unfortunately leads to a high false positive rates for matches because it is concerned with finding
the 'best' match according to this unidimensional criteria, where as our matching problem is actually
a multicriterion matching problem and we are more interested in accurately choosing 'obvious' matches
and flagging the rest for review.

The new MatchGraph implementation uses a similar criteria to find the 'best' match for each source instance,
but _only_ accepts the match if it is also a dominant solution: if any other matches are better matches
on any of the individual criterions, it is not automatically matched and is flagged for review.