AI-880 Fixed dashboard sharing in reports dialog
To place a report on user's dashboard, the report must be shared.
However, (and this is probably a long-standing issue), if you choose
"Dashboard" without "Share", then the report is not shared and does not
appear on the user's dashboard.

This commits add the behavior to automatically check shared if dashboard
is slected.