AI-1210 Migrated landing/login/signup pages to Bootstrap 3
This changes is necessary to be able to launch the new AI.org website which
includes a theme based on bootstrap 3.

This commit also removes the ActivityInfo.org specific theme: we will now provide the
custom scaffolding theme at deploy time so that the "about" pages can be updated
independently.