Configuration fixes for instrumented acceptance tests
To obtain code coverage, we instrument the development server with jacoco.
Apparently, this increases the amount of permgen space needed by the
development server.

This can also be fixed by adding -XX:MaxPermSize=2048m to the
JAVA_TOOL_OPTIONS environment variable, but I'm adding it here to avoid
future problems.