AI-944 Fixed admin pivoting and added caching for admin levels
Currently, we allow users to change location type of an activity, which
means that a location field may actually reference other location types
that one explicitly specified.

Since the query engine relies on the reference field to determine joins,
we need to include these old location types in the field type definition.