AI-944 Fixed unit test
The test fixture is missing reporting period entries for a few sites with
reportingfrequency=0, which led the old engine to ignore them. But the new
engine does not need to consult the reporting periods and so gets the correct
answer.

I don't want to update the test fixtures until everything else is passing and I
can verify other tests aren't affected by the corrected fixtures.