AI-944 Fixed drill down when sites do not have locations
The new data model is taking the next step of removing nullary location types:
some forms will simply not have a location field so the location dimension category
for these sites will be missing.

Updated the acceptance test to accept a blank location value in place of the
"RDC" label.