Fixed race condition in ExprFunctions
This is probably quite rarely occurs, but I did see it happen when running acceptance tests
locally. A second threads apparently reached the lookupMap == null branch after it was set
to an empty hash map, but before it was fully initialized. The Initialization-on-demand holder idiom
solves this race condition without locking.