AI-944 Exposed activity metadata in mysql store to PivotAdapter
Ultimately we need to be able to express all metadata in the legacy model
(aggregation, activity category, database name, etc) in the terms of the new
model, but I'm not comfortable with making changes to the new model just yet.

In the meantime, this keeps PivotAdapter from depending on hibernate, and
avoids fetching the same metadata from mysql twice.