AI-228 Added language switch to user interface
This commit integrates import changes from rebar-appcahe:
Namely, rebar-appcache now provides a distinct URL for bootstrap scripts and
appcache manifests for each compiled language. This allows users to switch between
languages without having to wait for the appcache to be invalidated.

Additionally, we add a user interface element that allows the user to switch language.

The only remaining item is the ability to change the user's "preferred language" which
governs the language of emails and notifications, as well as the language used when no
locale is specified.

The relationship between locale and offline mode is complicated and may still need some
work.