AI-228 Fixed arabic date formats
The arabic date formats supplied by CDLR are meant to be used in the context of
an RTL document.

Since we are embedding Arabic text in an overall LTR document, we need to add
bidi unicode control escapes to ensure that the date is rendered RTL despite the
presence of western digits on the boundary of the strings.