AI-228 Fixes broken parsing of arabic lat/lng coordinates
Unfortunately, GWT's emulated version of Character.isDigit() *does not*
behave in the same way as the JRE's Character.isDigit(). The emulated version
ONLY checks for ASCII digits ("Western Arabic Numerals"),
not other digits such as Eastern Arabic numerals.

Even more unfortunately,  the emulated versions are *not* used during
"Development Mode", so the difference is only observable once everything
is compiled to Javascript.

This commit updates the parser to use it's own locale-sensitive version of
isDigit().