Subform model/ui refactor (wip)
- Simplify the modelling of subforms with a single "SubFormKind" enum.
- Simplify the UI by breaking into discrete components: RepeatingSubFormPanel and PeriodSubFormPanel

I've also eliminated the separation between a few "Presenters" and "Views": I think this pattern is really
only worth the added verbosity and complexity if the user interface logic really needs to be exhaustively
tested. And in most cases, seperating presentation from logic is quite difficult, and you see things like
HTML composition leaking into the presenter.

The FormDesigner is now working but not quite the data entry