Refactored HRD store to use Objectify library.
I had originally done the mapping between our classes and AppEngine entities
by hand, but the code was ugly and hard to understand. With Objectify, there is less
code and clearer to understand what is being stored where.

This does cause breaking changes in the format entities are
stored in the datastore.