AI-1440: Batch ALL queries from a pivot table into a single batch
For example, if a pivot table includes indicator values from 100 forms,
this was previously 100 independent queries. If each of these 100 forms
was joined to provinces and territories, it would have neccessitated the
province table back and forth from memcache 100 times.

In addition, we are now ensuring that each individual join is performed
once.