AI-1498 - Fix corrupted partner fields
It appears there are a number of forms with partner fields serialized pointing to the wrong database.
I'm not sure if this is the result of a manual change that we (incompletely) applied to the database on behalf
of the user or if there was a past bug that leads to corrupted state. But this should patch all forms on load.