AI-1553 Fixes CloneDatabaseHandler
When TextType was "promoted" to a ParametrizedType when input
masks were added, it was no longer handled by the case.

This commit refactors the code to use the visitor pattern, so then as
new field types are added, the compiler can verify that a case
is properly added for them.