AI-1590 Do not accept malformed form ids
Previously, because of the way lookup was done, the form id "a1" worked as well
as "a0000000001". However, accepting these ids lead to internal inconsistencies and
errors in many places because they are not treated as equal by most code (and shouldn't be).

The API will now no longer accept malformed form ids - a query for form "a1" will return 404