AI-1602: For expressions that match multiple nodes in the tree, use the closest to the base table
For example, a selector like [Form Id].[Field] might match multiple nodes in the tree if multiple
forms reference [Form Id]. This happens with references to the partner form because nearly all
forms reference it.

Instead of throwing an error, we will just use the closest partner reference to the base table.

The formula syntax would allow selection of a more specific partner reference (e.g. Parent.partner.label)
but we need to expand the UI to allow for this.