AI-1651 Updated algorithm for default form query
Previously, the default query was assembled by finding all leaf fields
in the tree and adding them to the query with their full path. So for
classic forms, this could result in mutliple references to the same field,
for example:

* location.admin.parent.parent.parent.parent.name: Nord Kivu
* location.admin.parent.name: Nord Kivu
* location.admin.name: Nord Kivu

depending on whether the location field referenced a province, territoire, or lower-level name
record.

This updates the algorithm to the following:

1. Add the root record id
2. add all root leaf fields, with the alias "{field}"
3. do a breadth-first search for ALL referenced forms
3.b. For each referenced form found, add all leaf fields with the column id in the format {form}.{field}

So instead of the mess above, you will get:

@id:
indicator1,
indicator2,
indicator3,
partner.name
project.name,
localitie.name,
province.name,
territoire.name,

etc.