Refactored json library for compatibility with old GWT dev mode
- This change also leads to more efficient/correct compiled javascript. I'm not
actually sure why the previous implementation worked at all, to be honest.

Basically, our goal is the code executing on the client (browser) side should be able to parse
incoming JSON using JSON.parse() and then immediately work with the resulting value without
having to build it up into a tree of additional objects.

However, this means that we really can't use the Java "instanceof" operator. This would require
an additional step to decorate the entire JSON tree with additional Javascript objects holding
the corresponding typeid.

So the only way to do this is really to give up on polymorphism all together.

We're left with a single JsonValue interface, whose "type" can be queried by method calls
like "isObject" or "isBoolean" but not via the Java _instanceof_ operator.

This ensures that we can use the same interface on the server side and the browser side.