AI-1774 Hot-fix for Turkish users
A regression in String.toUpperCase() in GWT 2.8.0 now prevents users
whose operating system is set to the Turkish locale from loading AI:
https://github.com/gwtproject/gwt/issues/9540

As we use String.toUpperCase() to parse the lower case JSON string "single"
to the uppercase Cardinality.SINGLE.

This regression is fixed in GWT 2.8.2, but upgrading to GWT 2.8.2 appears
to require a bit more work to deal with changes to the GWT compiler anticipating
Java 9.

In the meantime, I've deployed a patched build of GWT 2.8.0 with the fix
for the above regression to nexus.bedatadriven.com:
https://github.com/bedatadriven/gwt/tree/v2.8.0-patched

And updated here.