AI-1811 Fix issue save/loading zero values in Json
GWT while compile the expression "object == null" to simply "!object".
When a JsValue instance points to a Javascript primitive, then the Java
expression "number == null" compiles to "!number" while evaluates to TRUE
when number === 0.

I will work on a patch to the GWT compiler, but in the meantime, avoid any null
checks on JsonValues using the Java syntax "jsonValue == null" or "jsonValue != null".
Instead, avoid null references to JsonValues and rely on JsonValue.isNull().