Removing appengine-export as we now can access java.awt
When we originally moved to AppEngine in 2012, we needed to patch
a number of the libraries we were using for exporting to reports as
the AppEngine runtime didn't allow access to java.awt or javax.imageio.

These were hosted at:
https://github.com/bedatadriven/appengine-export

With the move to the AppEngine java8 runtime, these restrictions has
been droped and we can migrate back to the real libraries.