Collapse gwt permutations to prevent permutation explosion
With an increasing number of languages, building a separate js app
for each browser/language pair is not feasible.

GWT does allow us to collapse certain permutations as "soft permutations"
that are switched at runtime, but our app cache linker wasn't able to
support this.

For this reason, I'm also migrating to an external "gwt-appcache"
project that provides similar functionality but has been more
regularly maintained. It's also based heavily on our own rebar-appcache
module so most changes are minimal.