Fixes for soft-deleted UserPermission records - When a User has _never_ been added to a Database, there exists no UserPermission record for that given [UserId,DatabaseId] and therefore no corresponding DatabaseGrant - When a User is added to a Database, and assigned any set of valid permissions, there exists a UserPermission record  and therefore a corresponding DatabaseGrant - When a User is removed from a Database, their UserPermissions record is _not_ removed but "soft-deleted" by setting all permissions on that record to FALSE and updating the version.
Under the new implementation, we need to continue to generate a DatabaseGrant for these soft-deleted UserPermission records. This is because the Synchronisation mechanism requires a comparison of local and remote versions of the User's Database version and UserPermission version.

If a DatabaseGrant record is not generated, the subsequent UserDatabaseMeta version defaults to the DatabaseMeta version (equivalent to the Database version) which remains unchanged by updates to a specific user's UserPermission record. Therefore the sync mechanism fails to update correctly, as it sees both local and remote versions as being the same.