Update guava library to 24.1.1-jre
Previous version of Guava (23.0, 15.0) had the known vulnerability
CVE-2018-10237:

Unbounded memory allocation allows remote attackers to conduct denial
of service attacks against servers that depend on this library and
deserialize attacker-provided data, because the AtomicDoubleArray
class (when serialized with Java serialization) and the
CompoundOrdering class (when serialized with GWT serialization)
perform eager allocation without appropriate checks on what a client has
sent and whether the data size is reasonable.

ActivityInfo does not use java deserialization on untrusted data,
but does include CompoundOrdering in the GWT RPC whitelist
and so would have been vulnerable to a DoS attack.