Update Google http client version
com.google.http-client 1.23.0 depended on httpclient-4.0.1, which
has two known vulnerabilities:

CVE-2011-1498: Apache HttpClient 4.x before 4.1.1 in Apache
HttpComponents, when used with an authenticating proxy server, sends the
Proxy-Authorization header to the origin server, which allows remote web
servers to obtain sensitive information by logging this header.

CVE-2014-3577: org.apache.http.conn.ssl.AbstractVerifier in Apache
HttpComponents HttpClient before 4.3.5 and HttpAsyncClient before 4.0.2
does not properly verify that the server hostname matches a domain name
in the subject's Common Name (CN) or subjectAltName field of the X.509
certificate, which allows man-in-the-middle attackers to spoof SSL
servers via a "CN=" string in a field in the distinguished name (DN) of
a certificate, as demonstrated by the "foo,CN=www.apache.org" string
in the O field.

As this library was only used in production to communicate between
the application server and Google APIs over Google's internal network,
it is unlikely that these vulnerabilities could have been exploited.