HOTFIX AI-2029 - Do not block sync on deletion of deleted records
This introduces the idea of transaction validation modes. The default
mode is STRICT, in which a transaction will fail if it is in any way
inconsistent with the current form schema.

For example, if the update includes a change to a deleted or non-existant
field, or a change to a form that has been deleted, or even a deletion
of a record that has already been deleted, it will fail.

This is the kind of feedback we want to provide to API clients, so that
we don't silently ignore real problems in the API usage.

However, when a transaction is composed while offline, there is the
chance of drift between the schema against which it was composed, and
the schema of a form at the time of synchronization.

Some differences are genuine conflicts which we cannot ignore. For
example, if the user's permissions change while offline, and they
no longer have permission to submit 250 new records, then we should
block synchronization until the user's permissions are restored, or
the user elects to discard their offline data.

Other differences can be (more or less) safely ignored. For example,
if a user deletes a record while offline, and then the same record
is deleted by someone else (or the same user with a different device)
before synchronization, the conflict can be safely and silently ignored.

With the OFFLINE transaction mode, a transaction with this class
of conflict will still be accepted, and synchronization can continue.