Upgrade to latest version of Jackson2
Jackson 2.9.5 is a transitive dependency of the swagger library was
flagged with the following vulnerabilities:
* CVE-2018-14719
* CVE-2018-14720
* CVE-2018-14721
* CVE-2018-1000873

All of these vulnerabilities affect the Jackson library only when
accepted untrusted input (XML or JSON) from remote attackers.

Swagger and Jackson2 deserialization are only used at build
time to generate static API documentation and has never been
deployed with the ActivityInfo server.

Jackson2 annotations are also used to mark required fields that
are required. This is a seperate dependency that does not include
the affected deserialization library.

Based on this analysis, ActivityInfo was not affected by these
vulnerabilities, but we are upgrading the dependencies out of an
abundance of caution.