Fix cloning of UserPermissions associated with Default Partner
Previously when cloning a database incl. partners and user permissions, the CloneDatabaseHandler:
- Created a new Default Partner for the Database.
- Copied all Partners from the original Database, *except* its Default Partner.
- Copied all UserPermissions from the original Database.

The CloneDatabaseHandler *did not* update the Partner references on UserPermissions in the new Database.
Therefore any users with Default partner references will still (incorrectly) reference the Default partner in the _old_ Database.

Fix for this issue ensures that any references to the old Default Partner are mapped to the new Default Partner.

# Conflicts:
#	server/src/main/java/org/activityinfo/server/command/handler/CloneDatabaseHandler.java