AI-2052 Fix bug with reordering of fields in keyed subforms
There exists a discrepancy between the Form Element index and the Widget index for keyed subforms (Monthly,Daily, etc.). The period field on a keyed subform is included on the form element list, but is *not* included as a Widget on the subform drop panel (this is handled specially elsewhere).

Therefore, when we simply use the widgetIndex when reinserting the element via FormClass::insertElement, we are in fact offset by -1 from the correct Form Element index. This led to the experience of not being able to move a form field to the very bottom of a subform in the form designer, or the field appearing above where it was intended.

Further, when a field is moved to the top of a subform, it was able to move the index of the period key field itself (as the inserted element was index 0 and pushed the period field down the list). As the issue went unnoticed and unresolved for a significant period of time, an unknown number of form schema have been affected and we must assume the period field could now reside anywhere on a form element list.

Therefore, we find the correct form element index by incrementing the widgetIndex by +1 *IF*:
 * The form is a keyed subform
 * The index of the period key form element is less than or equal to the current widgetIndex